Moving Web
Added a nav entry for moving web and Conways game of life
added moving web page with a project that looks like a bunch of nodes of a web moving across the screen and attatching to eachother.
removed old services.html adn replaced with web. at this point most of the projects are in their own page. i may bring back the liquid as its own page but maybe not.